pyproject.toml
[project]
name = "virtu-escpos-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.115.5",
    "python-escpos[all]>=3.1",
]
